pyproject.toml
[project]
name = "flask-user-management"
version = "1.0.0"
description = "Flask application with user authentication and CRUD operations"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-mysqldb>=2.0.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
